# Avoid adding empty splices in `CodedBufferWriter`